[experiments] Add abstraction for configcat to work in self-hosted
