+++ components/gitpod-protocol/package.json
@@ -42,6 +42,7 @@
         "@types/react": "17.0.32",
         "ajv": "^6.5.4",
         "analytics-node": "^6.0.0",
+        "configcat-node": "^7.0.0",
         "cookie": "^0.4.2",
         "express": "^4.17.3",
         "inversify": "^5.1.1",
