+++ components/gitpod-protocol/src/experiments/types.ts
@@ -1,0 +1,32 @@
+/**
+ * Copyright (c) 2022 Gitpod GmbH. All rights reserved.
+ * Licensed under the GNU Affero General Public License (AGPL).
+ * See License-AGPL.txt in the project root for license information.
+ */
+
+import { Team } from "../teams-projects-protocol";
+
+// Attributes define attributes which can be used to segment audiences.
+// Set the attributes which you want to use to group audiences into.
+export interface Attributes {
+    userId?: string;
+    email?: string;
+
+    // Currently selected Gitpod Project ID
+    projectId?: string;
+
+    // Currently selected Gitpod Team ID
+    teamId?: string;
+    // Currently selected Gitpod Team Name
+    teamName?: string;
+
+    // All the Gitpod Teams that the user is a member (or owner) of
+    teams?: Array<Team>;
+}
+
+export interface Client {
+    getValueAsync<T>(experimentName: string, defaultValue: T, attributes: Attributes): Promise<T>;
+
+    // dispose will dispose of the client, no longer retrieving flags
+    dispose(): void;
+}
